+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/) and this
 project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [0.3.0] - Unreleased
+### Added
+- Allow to select the menu style type.
+- Allow to customize the basename of the legal pages.
+
 ### Changed
 - Upgrade to Lume 3.